first basic chat version with express
a new user has to register with the website first by providing an unique
user name (X - aber noch kein visual feedback)
in the chat room, all messages should be displayed in chronological
order with a time stamp and the user name of the user who has posted the
message. (X)
a marked message is generated and sent to the chat room each time a new
user connects to the chat room or leaves that chat room. (noch nicht
"marked")
a list of all user in the chat room is displayed. (bisher nur im
console.log)
a user can send a message to any one dedicated recipient (private
message) instead of to all online users in the chat room (fehlt noch)
a user can send multi-media files to the participants of a chat room or
via a private message, e.g., pictures, movies, and sound files. Files
must not be stored on the server! (fehlt noch)